Changed Card Number from Int to Long

+++ Backend/user-service/src/main/java/UserService/UserService/Controller/UserController.java
@@ -63,7 +63,7 @@ public class UserController {
         String address = body.get("address");
         boolean notifications = Boolean.parseBoolean(body.get("notifications"));
         String cardName = body.get("cardName");
-        int cardNumber = Integer.parseInt(body.get("cardNumber"));
+        Long cardNumber = Long.parseLong(body.get("cardNumber"));
         String cardExpiration = body.get("cardExpiration");
         int cardCVV = Integer.parseInt(body.get("cardCVV"));
