+++ Backend/user-service/src/main/java/UserService/UserService/Model/User.java
@@ -2,7 +2,7 @@ package UserService.UserService.Model;
 
 public record User(
     Integer userID, String firstname, String lastname, String username, String password, String email,
-    String address, boolean notifications, String cardName, Integer cardNumber, String cardExpiration, Integer cardCVV
+    String address, boolean notifications, String cardName, Long cardNumber, String cardExpiration, Integer cardCVV
 ) 
 
 {}
